Basic network communication
Got connect.cpp running on Duino to talk to app running on phone. This
required fixing sock.cpp's way of sending messages on UDP (connect
should not be used, was the library made for TCP?). Added some debugging
print statements to main to work out that it crashes when creating the
listener thread on the Duino. Project settings possibly modified to have
intellisense for c++ 11, and VS sln probably modified because I opened
the project in VS.